+++ CHANGES.md
@@ -16,6 +16,7 @@ Version 1.14 [????.??.??]
   to two arguments, just like `InfixT` is desugared. Similarly, attempting to
   desugar a `PromotedUInfixT` results in an error, just like attempting to
   desugar a `UInfixT` would be.
+* `th-desugar` now supports `DefaultD` (i.e., `default` declarations).
 * Fix an inconsistency which caused non-exhaustive `case` expressions to be
   desugared into uses of `EmptyCase`. Non-exhaustive `case` expressions are now
   desugared into code that throws a "`Non-exhaustive patterns in...`" error at